+++ src/com/abstraction/Rectangle.java
@@ -1,0 +1,17 @@
+package com.abstraction;
+
+public class Rectangle implements Shape {
+	private double length;
+    private double width;
+     
+    public Rectangle(double length, double width) { 
+       this.length = length; 
+       this.width = width; 
+    } 
+    public void draw() { 
+       System.out.println("Drawing Rectangle"); 
+    } 
+    public double getArea() { 
+       return length*width; 
+    } 
+}
